Fix OpenStack volume creation [SENTRY-878]

+++ src/waldur_core/structure/__init__.py
@@ -495,8 +495,8 @@ class ServiceBackend:
 
     @staticmethod
     def mb2gb(val):
-        return val / 1024 if val else 0
+        return int(val / 1024) if val else 0
 
     @staticmethod
     def mb2tb(val):
-        return val / 1024 / 1024 if val else 0
+        return int(val / 1024 / 1024) if val else 0
